aruha-400 remove Nullables by separating EventType and EventTypeBase
We have a very strict rule for null handling in Nakadi. In order to
avoid Nullable fields, we split EventType class in two to represent two
different moments of this object: when some fields are abset (base) and
after all fields are defined (EventType).
