+++ src/main/java/org/zalando/nakadi/enrichment/Enrichment.java
@@ -1,13 +1,14 @@
 package org.zalando.nakadi.enrichment;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 import org.zalando.nakadi.domain.BatchItem;
 import org.zalando.nakadi.domain.EnrichmentStrategyDescriptor;
 import org.zalando.nakadi.domain.EventCategory;
 import org.zalando.nakadi.domain.EventType;
+import org.zalando.nakadi.domain.EventTypeBase;
 import org.zalando.nakadi.exceptions.EnrichmentException;
 import org.zalando.nakadi.exceptions.InvalidEventTypeException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -22,7 +23,7 @@ public class Enrichment {
         this.registry = registry;
     }
 
-    public void validate(final EventType eventType) throws InvalidEventTypeException {
+    public void validate(final EventTypeBase eventType) throws InvalidEventTypeException {
         if (eventType.getCategory() == EventCategory.UNDEFINED && !eventType.getEnrichmentStrategies().isEmpty()) {
             throw new InvalidEventTypeException("must not define enrichment strategy for undefined event type");
         }
