+++ src/main/java/org/zalando/nakadi/repository/EventTypeRepository.java
@@ -1,6 +1,7 @@
 package org.zalando.nakadi.repository;
 
 import org.zalando.nakadi.domain.EventType;
+import org.zalando.nakadi.domain.EventTypeBase;
 import org.zalando.nakadi.exceptions.DuplicatedEventTypeNameException;
 import org.zalando.nakadi.exceptions.InternalNakadiException;
 import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
@@ -10,7 +11,7 @@ import java.util.Optional;
 
 public interface EventTypeRepository {
 
-    void saveEventType(EventType eventType) throws InternalNakadiException, DuplicatedEventTypeNameException;
+    EventType saveEventType(EventTypeBase eventType) throws InternalNakadiException, DuplicatedEventTypeNameException;
 
     EventType findByName(String name) throws InternalNakadiException, NoSuchEventTypeException;
