+++ src/main/java/org/zalando/nakadi/validation/schema/SchemaEvolutionConstraint.java
@@ -1,9 +1,10 @@
 package org.zalando.nakadi.validation.schema;
 
 import org.zalando.nakadi.domain.EventType;
+import org.zalando.nakadi.domain.EventTypeBase;
 
 import java.util.Optional;
 
 public interface SchemaEvolutionConstraint {
-    Optional<SchemaEvolutionIncompatibility> validate(EventType original, EventType eventType);
+    Optional<SchemaEvolutionIncompatibility> validate(EventType original, EventTypeBase eventType);
 }
